move splitmerge from /pkg/storages to /internal
