+++ internal/splitmerge/channelreader.go
