+++ internal/splitmerge/channelwriter.go
