+++ internal/splitmerge/fixedsizewriter.go
