+++ internal/splitmerge/splitreader.go
@@ -1,8 +1,9 @@
 package splitmerge
 
 import (
-	"github.com/wal-g/tracelog"
 	"io"
+
+	"github.com/wal-g/tracelog"
 )
 
 func SplitReader(reader io.Reader, parts int, blockSize int) []io.Reader {
@@ -10,7 +11,7 @@ func SplitReader(reader io.Reader, parts int, blockSize int) []io.Reader {
 	channels := make([]chan []byte, 0)
 
 	for i := 0; i < parts; i++ {
-		channels = append(channels, make(chan []byte, 0))
+		channels = append(channels, make(chan []byte))
 		result = append(result, NewChannelReader(channels[i]))
 	}
