+++ internal/stream_fetch_helper.go
@@ -6,7 +6,7 @@ import (
 	"path"
 	"time"
 
-	"github.com/wal-g/wal-g/pkg/storages/splitmerge"
+	"github.com/wal-g/wal-g/internal/splitmerge"
 
 	"github.com/wal-g/tracelog"
 	"github.com/wal-g/wal-g/internal/compression"
