+++ internal/stream_push_helper.go
@@ -8,7 +8,7 @@ import (
 
 	"golang.org/x/sync/errgroup"
 
-	"github.com/wal-g/wal-g/pkg/storages/splitmerge"
+	"github.com/wal-g/wal-g/internal/splitmerge"
 
 	"github.com/wal-g/tracelog"
 	"github.com/wal-g/wal-g/utility"
